fastApiThird/pyproject.toml
[project]
name = "fastapithird"
version = "0.1.0"
description = ""
authors = [
    {name = "Melve",email = "shvechkovmaxcanada@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standart] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose[cryptografy] (>=3.4.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.4.0,<9.0.0)"
]

[tool.poetry]
packages = [{include = "fastapithird", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
